Ready to prod

+++ backend/src/main/resources/db/migration/V1__created_table_product.sql
@@ -1,7 +1,7 @@
 CREATE TABLE tb_product (
     id SERIAL PRIMARY KEY,
     name varchar(255),
-    price DOUBLE,
+    price DOUBLE PRECISION,
     image_Uri varchar(255),
     description varchar(255)
 );
