+++ backend/src/main/resources/db/migration/V2__created_table_order.sql
@@ -1,8 +1,8 @@
 CREATE TABLE tb_order (
     id SERIAL PRIMARY KEY,
     address varchar(255),
-    latitude DOUBLE,
-    longitude DOUBLE,
+    latitude DOUBLE PRECISION,
+    longitude DOUBLE PRECISION,
     moment TIMESTAMP,
     status SMALLINT
 );
